Updated all scripts.
In particular, the 'mustang-cuts' script has been finalized. I has been
written to run on SciNet's Niagara, and includes a sample shell script
which can be used for job submissions in that cluster.

The 'data-selection' and 'xor' scripts have undergone minor formatting
changes and typo corrections which do not affect their functioning.